backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
python-dotenv==1.0.1
python-multipart==0.0.12
openpyxl==3.1.5
httpx==0.27.2
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
alembic==1.13.3
sentry-sdk[fastapi]==2.14.0
